MediaService/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

# Fetch all the keys stored in the env
DATABASE_URL = os.getenv("DATABASE_URL")

# AWS configuration keys
S3_BUCKET = os.getenv("S3_BUCKET")
AWS_ACCESS_KEY_ID = os.getenv("AWS_ACCESS_KEY_ID")
AWS_SECRET_ACCESS_KEY = os.getenv("AWS_SECRET_ACCESS_KEY")
AWS_REGION = os.getenv("AWS_REGION")

JWT_SECRET = os.getenv("JWT_SECRET")
JWT_ALGORITHM = "HS256"
