MediaService/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
boto3==1.34.21
python-jose==3.3.0
